Update QFace.Sdk.AI project with new AI libraries and refactor algorithms
- Added `Microsoft.Extensions.AI` and related packages for enhanced AI capabilities.
- Refactored algorithm files to remove unused using directives, improving code cleanliness.
- Updated `AnthropicProvider` and `GoogleGeminiProvider` to use the new Microsoft.Extensions.AI abstractions, with stubs for future implementation.
- Enhanced `OpenAIProvider` to utilize the new chat client structure and improved logging.
- Adjusted service methods to streamline forecasting and skills analysis processes.

+++ src/QFace.Sdk.AI/Algorithms/RegressionAnalysisAlgorithm.cs
@@ -1,6 +1,4 @@
 using MathNet.Numerics.LinearRegression;
-using Microsoft.Extensions.Logging;
-using QFace.Sdk.AI.Models;
 
 namespace QFace.Sdk.AI.Algorithms;
 
@@ -55,7 +53,7 @@ public class RegressionAnalysisAlgorithm : IForecastAlgorithm
         var residuals = yValues.Select((y, i) => y - (intercept + slope * xValues[i])).ToArray();
         var residualVariance = residuals.Sum(r => r * r) / (dataCount - 2);
         var standardError = Math.Sqrt(residualVariance);
-        var zScore = GetZScore(confidenceLevel);
+        var zScore = (double)GetZScore(confidenceLevel);
         var margin = (decimal)(standardError * zScore);
 
         var result = new ForecastResult
